refactor: file forms utils
