+++ src/lib/forms/index.ts
@@ -2,5 +2,6 @@ import Form, { createForm } from './form.svelte';
 import FieldController from './fieldController.svelte';
 import FormError from './formError.svelte';
 import Input from './input.svelte';
+import { zodFile } from './utils';
 
-export { Form, createForm, FieldController, FormError, Input };
+export { Form, createForm, FieldController, FormError, Input, zodFile };
